cast to int if rounded down to int

+++ PyPoE/poe/file/translations.py
@@ -380,9 +380,9 @@ class TranslationQuantifier(object):
         'old_leech_percent': lambda v: v/5,
         'old_leech_permyriad': lambda v: v/50,
         # TODO dp = precision?
-        'per_minute_to_per_second_0dp': lambda v: round(v/60, 0),
+        'per_minute_to_per_second_0dp': lambda v: int(round(v/60, 0)),
         'per_minute_to_per_second_2dp': lambda v: round(v/60, 2),
-        'milliseconds_to_seconds_0dp': lambda v: round(v/1000, 0),
+        'milliseconds_to_seconds_0dp': lambda v: int(round(v/1000, 0)),
         'milliseconds_to_seconds_2dp': lambda v: round(v/1000, 2),
         # Only once TODO
         #'multiplicative_damage_modifier': lambda v: v,
